% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-lex.R
\docType{data}
\name{asm_patterns}
\alias{asm_patterns}
\title{Regex patterns for parsing 6502 assembly}
\format{
An object of class \code{character} of length 24.
}
\usage{
asm_patterns
}
\description{
Regex patterns for parsing 6502 assembly
}
\keyword{datasets}
