% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003-compile.R
\name{compile}
\alias{compile}
\title{Compile ASM to a c64 exectuable (in PRG format)}
\usage{
compile(asm, prg = TRUE)
}
\arguments{
\item{asm}{single character string containing all ASM (including newlines)}

\item{prg}{Should compilation return the bytes of the compiled PRG?
Default: TRUE.  If FALSE, return the main assembler data.frame which 
is more useful for debugging.}
}
\value{
prg bytes as a raw vector or \code{prg_df} main assembler data.frame
}
\description{
Compile ASM to a c64 exectuable (in PRG format)
}
