% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-lex.R
\name{create_line_tokens}
\alias{create_line_tokens}
\title{Split each ASM line into tokens}
\usage{
create_line_tokens(asm)
}
\arguments{
\item{asm}{a single character string containing full ASM code, including newlines}
}
\value{
A list where element contains the tokens for a single row.
        Rows with just comments or whitespace are discarded
}
\description{
Split each ASM line into tokens
}
