% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-parse.R
\name{create_prg_df}
\alias{create_prg_df}
\title{Create the main compiler datastructure 'prg_df' from a list of `line_tokens`}
\usage{
create_prg_df(line_tokens)
}
\arguments{
\item{line_tokens}{list of line tokens created by `create_line_tokens()`}
}
\value{
prg_df the main compiler datastructure
}
\description{
Create the main compiler datastructure 'prg_df' from a list of `line_tokens`
}
