% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-parse.R
\name{create_prg_df_row}
\alias{create_prg_df_row}
\title{Determine the opcode, address mode etc for a sequence of tokens representing a single instruction}
\usage{
create_prg_df_row(tokens)
}
\arguments{
\item{tokens}{A sequence of ASM tokens for one row of the ASM file.}
}
\value{
single row data.frame containing information about the row
}
\description{
Determine the opcode, address mode etc for a sequence of tokens representing a single instruction
}
