% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002-process.R
\name{extract_prg_bytes}
\alias{extract_prg_bytes}
\title{Extract the raw bytes for the assembled PRG}
\usage{
extract_prg_bytes(prg_df)
}
\arguments{
\item{prg_df}{main assembler data.frame}
}
\value{
raw vector of bytes
}
\description{
Extract the raw bytes for the assembled PRG
}
