% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-parse.R
\name{get_opcode_info}
\alias{get_opcode_info}
\title{Determine the opcode, address mode etc for a sequence of tokens representing a single instruction}
\usage{
get_opcode_info(tokens)
}
\arguments{
\item{tokens}{A sequence of ASM tokens. This must start with a valid opcode, and
not be a `symbol` or an assembler directive like `.text`}
}
\value{
single row data.frame containing information about the instruction
}
\description{
Determine the opcode, address mode etc for a sequence of tokens representing a single instruction
}
