% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999-utils.R
\name{h2i}
\alias{h2i}
\title{Hex (character strings) to integer values}
\usage{
h2i(h)
}
\arguments{
\item{h}{character vector of values e.g. "#$1a", "$12", "#127", "#b00111111"}
}
\value{
vector of integers
}
\description{
Any leading '#' will be dropped.
Values prefixed with "$" will be treated as hexadecimal.
Values with a leading 'b' will be treated as binary (MSB bit first)
}
