% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999-utils.R
\name{hi}
\alias{hi}
\title{Fetch the high byte of a 16bit address}
\usage{
hi(address)
}
\arguments{
\item{address}{address from $0000 to $ffff}
}
\value{
integer value for the high byte of this address
}
\description{
Fetch the high byte of a 16bit address
}
