% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999-utils.R
\name{lo}
\alias{lo}
\title{Fetch the low byte of a 16bit address}
\usage{
lo(address)
}
\arguments{
\item{address}{address from $0000 to $ffff}
}
\value{
integer value for the low byte of this address
}
\description{
Fetch the low byte of a 16bit address
}
