% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999-utils.R
\docType{data}
\name{opcodes}
\alias{opcodes}
\title{Reference list of opcode information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 217 rows and 15 columns.
}
\source{
\url{http://www.oxyron.de/html/opcodes02.html}
}
\usage{
opcodes
}
\description{
A dataset containing opcode information for 6502 instructions
}
\keyword{datasets}
