% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002-process.R
\name{process_symbols}
\alias{process_symbols}
\title{Process any cross-referenced symbols to insert actual addresse/values}
\usage{
process_symbols(prg_df)
}
\arguments{
\item{prg_df}{main assembler data.frame created by `create_prg_df()`}
}
\value{
udpated prg_df with actual addresses added for each row, and
        cross-referenced values resolved
}
\description{
Process any cross-referenced symbols to insert actual addresse/values
}
