% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002-process.R
\name{process_zero_padding}
\alias{process_zero_padding}
\title{Pad out with zero bytes between address blocks}
\usage{
process_zero_padding(prg_df)
}
\arguments{
\item{prg_df}{main assembler data.frame after `process_symbols()` has been run}
}
\value{
prg_df updated with zero padding rows between address blocks so that
        every byte from start to finish is accounted for
}
\description{
Pad out with zero bytes between address blocks
}
