% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999-utils.R
\name{s2i}
\alias{s2i}
\title{String to ASM for ".text" code}
\usage{
s2i(string)
}
\arguments{
\item{string}{single string}
}
\value{
Vector of integers
}
\description{
This is a complete hack to create CBMASCII bytes from plain text
}
